Refactor - moved ChromaOutrageShamanCount to shared area
